refactor: mover componentes a archivos dedicados
